backend/requirements.txt
boto3>=1.34
jinja2>=3.1
pyjwt[crypto]>=2.8
requests>=2.31
